Fix heap buffer overflow in setVersionInfo
I caught this bug running the connection code with the address sanitzier turned on.

`sizeof` returns the size of the object representation of the type argument at
compile time. This can differ from the actual size of `[stringData bytes]` depending on
the length of the string. If the string used to construct the data was smaller than the
value returned by `sizeof`, this memcopy will read beyond the `stringData` buffer and
copy potentially unitialized/garbage data into `appVer`.

`[stringData length]` will return the actual length of the buffer provided by
`[stringData bytes]`