Fix framework build for Xcode 12
Xcode 12 includes a slice for the "arm64" architecture when building
for the simulator (in addition to the usual "i386" and "x86_64"
architectures for Xcode <12). This is probably for supporting the
simulator on (future) Macs using Apple Silicon processors.

Since the device build also includes the "arm64" architecture, lipo does
not know which of the two "arm64" slices you want and refuses to create
a combined fat binary framework.

This results in the following error:

```Fatal error: lipo: .../Debug-iphonesimulator/... and .../Debug-iphoneos/...
Have the same architectures (arm64) and can't be in the same fat output file```

As a workaround, this commit excludes the "arm64" architecture from the
simulator build by appending the EXCLUDED_ARCHS build variable per:

https://stackoverflow.com/questions/64022291/